[XPU] Add WITH_XPU_CADA: run CINN on M100 via xtrans CUDA-compat
Introduces a WITH_XPU_CADA build option that lets CINN's existing
CUDA codegen/NVRTC/runtime path target Baidu KUNLUN M100 through
xtrans, a CUDA/NVRTC/Driver-API compatible toolchain, instead of
requiring a real NVIDIA CUDA install.

Key changes:
- CMakeLists.txt / cmake/cinn.cmake / cmake/cinn/*.cmake: new
  WITH_XPU_CADA option, xtrans toolchain root/library discovery
  (XTRANS_ROOT), and isolation of the handful of spots where xtrans's
  layout differs from a stock NVIDIA CUDA install (no lib64/stubs,
  0-byte NVTX stub, cusolver->cusparse link dependency, etc).
- paddle/cinn/backends/cuda_util.h, paddle/cinn/runtime/cuda/cuda_util.cc,
  paddle/phi/backends/dynload/cudnn.h: CINN's cudnn calls are ported to
  phi's existing dynload/lazy-dlopen wrapper instead of link-time
  linking libcudnn.so. This avoids pulling xtrans's own libLLVM-15.so
  (a transitive dependency of xtrans's libcudnn.so) into any process
  that also links CINN's statically-linked LLVM 12, which otherwise
  collides at the global llvm::cl::opt registry and aborts at startup
  (hit in eager_generator).
- paddle/cinn/backends/nvrtc/header_generator.cc: skip injecting
  jitify's JIT-safe stub headers under WITH_XPU_CADA (xtrans already
  provides a real, self-consistent stdlib via --include-path, and
  jitify's <limits>/<cmath> stubs conflict with it).
- paddle/cinn/backends/nvrtc/nvrtc_util.cc: propagate
  -DPADDLE_WITH_XPU_CADA into the NVRTC device-code compile itself
  (previously only visible to host-side C++), so the WITH_XPU_CADA-only
  code paths injected into runtime headers actually activate.
- paddle/cinn/runtime/cuda/float16_xpu_cada.h (new),
  paddle/cinn/runtime/cuda/bfloat16.h,
  paddle/cinn/runtime/cuda/cinn_cuda_runtime_source.cuh,
  paddle/cinn/backends/codegen_cuda_dev.cc: xtrans/clang's stricter
  overload resolution (vs nvcc) makes cinn::common::float16/bfloat16's
  implicit float+double conversions ambiguous in several call sites
  (shuffle intrinsics, cmath max/min promotion, generic struct
  shuffling). A SFINAE-based generic struct-shuffle overload is added
  for WITH_XPU_CADA; float16/bfloat16/fp8/welford/argidx reduce code
  paths that still hit the remaining ambiguity are temporarily scoped
  out for WITH_XPU_CADA (float32 reduce is unaffected and verified
  working end-to-end via CINN on M100/xtrans).
- paddle/cinn/runtime/cuda/xtrans_cublas_v2_compat.h (new),
  paddle/cinn/runtime/cuda/cublas_util.h: work around xtrans's
  cublas_v2.h textually rewriting several cublas symbols to _v2
  variants that xtrans's own libcublas.so does not export.
- paddle/ap/*: fix xtrans/clang not implicitly instantiating member
  function templates reached only through `if constexpr` branches
  inside lambdas (explicit instantiation added), plus narrowing-
  conversion and API-usage fixes required to compile under xtrans/clang.

Verified: WITH_XPU_CADA+CINN build completes, and
simple_add_cinn.py (fp32 elementwise add) passes end-to-end through
CINN's JIT/NVRTC path on M100 via xtrans.

+++ paddle/ap/include/code_gen/arg_source_maker.h
@@ -72,7 +72,8 @@ struct ArgSourceMaker {
       auto DoEachIndex = [&](std::size_t index) -> Ok {
         ADT_CHECK(index < inputs.size());
         const auto& bir_node = inputs.at(index);
-        NativeIrValueSource native_ir_value_source{input_idx};
+        NativeIrValueSource native_ir_value_source{
+            static_cast<int>(input_idx)};
         TensorSource tensor_source{native_ir_value_source};
         InTensorSource in_tensor_source{tensor_source};
         Pair pair{bir_node, in_tensor_source};
@@ -84,7 +85,8 @@ struct ArgSourceMaker {
         for (std::size_t i = start; i < end; ++i) {
           ADT_CHECK(i < inputs.size());
           const auto& bir_node = inputs.at(i);
-          PackedIrValueSource packed_ir_value_source{input_idx, i};
+          PackedIrValueSource packed_ir_value_source{
+              static_cast<int>(input_idx), static_cast<int>(i)};
           TensorSource tensor_source{packed_ir_value_source};
           InTensorSource in_tensor_source{tensor_source};
           Pair pair{bir_node, in_tensor_source};
@@ -121,7 +123,7 @@ struct ArgSourceMaker {
         ADT_CHECK(index < outputs.size());
         const auto& bir_node = outputs.at(index);
         OutTensorSource out_tensor_source{
-            TensorSource{NativeIrValueSource{output_idx}}};
+            TensorSource{NativeIrValueSource{static_cast<int>(output_idx)}}};
         ret.emplace_back(Pair{bir_node, out_tensor_source});
         ++output_idx;
         return adt::Ok{};
@@ -130,8 +132,8 @@ struct ArgSourceMaker {
         for (std::size_t i = start; i < end; ++i) {
           ADT_CHECK(i < outputs.size());
           const auto& bir_node = outputs.at(i);
-          OutTensorSource out_tensor_source{
-              TensorSource{PackedIrValueSource{output_idx, i}}};
+          OutTensorSource out_tensor_source{TensorSource{PackedIrValueSource{
+              static_cast<int>(output_idx), static_cast<int>(i)}}};
           ret.emplace_back(Pair{bir_node, out_tensor_source});
         }
         ++output_idx;
